Refactor backends
* Make Backend abstract
* Add RemoteBackend
* Remove unused backend methods
