+++ cli/dstack/api/backend.py
@@ -1,7 +1,7 @@
 from typing import Dict, List, Optional
 
-from dstack.backend import Backend
 from dstack.backend.aws import AwsBackend
+from dstack.backend.base import Backend, RemoteBackend
 from dstack.backend.local import LocalBackend
 from dstack.core.error import BackendError
 
@@ -16,11 +16,11 @@ def get_all_backends():
 
 
 def list_backends() -> List[Backend]:
-    return [backend for backend in get_all_backends() if backend.loaded()]
+    return [backend for backend in get_all_backends() if backend.loaded]
 
 
 def dict_backends() -> Dict[str, Backend]:
-    return {backend.NAME: backend for backend in list_backends()}
+    return {backend.name: backend for backend in list_backends()}
 
 
 def get_backend_by_name(name: str) -> Optional[Backend]:
@@ -30,7 +30,7 @@ def get_backend_by_name(name: str) -> Optional[Backend]:
     return backend
 
 
-def get_current_remote_backend() -> Backend:
+def get_current_remote_backend() -> RemoteBackend:
     return get_backend_by_name(DEFAULT_REMOTE)
