+++ cli/dstack/api/logs.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from dstack.backend import Backend
+from dstack.backend.base import Backend
 from dstack.core.job import JobHead
 from dstack.core.repo import RepoAddress
