+++ cli/dstack/backend/aws/logs.py
@@ -225,49 +225,3 @@ def poll_logs(
             return
         else:
             raise e
-
-
-def query_logs(
-    s3_client: BaseClient,
-    logs_client: BaseClient,
-    bucket_name: str,
-    repo_address: RepoAddress,
-    run_name: str,
-    start_time: int,
-    end_time: Optional[int],
-    next_token: Optional[str],
-    job_host_names: Dict[str, Optional[str]],
-    job_ports: Dict[str, Optional[List[int]]],
-    job_app_specs: Dict[str, Optional[List[AppSpec]]],
-) -> Tuple[
-    List[LogEvent],
-    Optional[str],
-    Dict[str, Optional[str]],
-    Dict[str, Optional[List[int]]],
-    Dict[str, Optional[List[AppSpec]]],
-]:
-    job_host_names = dict(job_host_names)
-    job_ports = dict(job_ports)
-    job_app_specs = dict(job_app_specs)
-    filter_logs_events_kwargs = _filter_logs_events_kwargs(
-        bucket_name, repo_address, run_name, start_time, end_time, next_token
-    )
-    response = logs_client.filter_log_events(**filter_logs_events_kwargs)
-    return (
-        [
-            _render_log_message(
-                s3_client,
-                bucket_name,
-                event,
-                repo_address,
-                job_host_names,
-                job_ports,
-                job_app_specs,
-            )
-            for event in response["events"]
-        ],
-        response.get("nextToken"),
-        job_host_names,
-        job_ports,
-        job_app_specs,
-    )
