+++ cli/dstack/backend/aws/runs.py
@@ -128,15 +128,3 @@ def get_run_heads(
             run = runs_by_id[run_id]
             _update_run(ec2_client, s3_client, bucket_name, run, job_head, include_request_heads)
     return sorted(list(runs_by_id.values()), key=lambda r: r.submitted_at, reverse=True)
-
-
-def list_run_heads(
-    ec2_client: BaseClient,
-    s3_client: BaseClient,
-    bucket_name: str,
-    repo_address: RepoAddress,
-    run_name: Optional[str],
-    include_request_heads: bool,
-) -> List[RunHead]:
-    job_heads = jobs.list_job_heads(s3_client, bucket_name, repo_address, run_name)
-    return get_run_heads(ec2_client, s3_client, bucket_name, job_heads, include_request_heads)
